NCAA March Madness Basketball Data

A dataset that contains each college basketball teams' performances and public picks in the March Madness competition from 2016 to 2024. Please keep in mind that we only have public picks in 2024 data.

Usage

march_madness_dataset

Format

A data frame with 512 rows and 7 variables.

YEAR

Ending year of the team's season

TEAMNO

Unique identifier for the team

TEAM

Division I college basketball team

SEED

Preliminary ranking for the March Madness tournament

ROUND

Farthest round the team made it in the tournament

POWER

Calculation of the team's strength relative to other teams in the tournament by Heatcheck

FINALS

the percentage of people who picked the team to win the game in the Finals

Source

https://github.com/ETC5523-2024/assignment-4-packages-and-shiny-apps-Vincent0153/tree/main/marchmadness/data